+++ app/src/main/java/woowacourse/movie/list/view/TheaterBottomSheetFragment.kt
@@ -11,6 +11,7 @@ import woowacourse.movie.databinding.FragmentTheaterBinding
 import woowacourse.movie.detail.view.MovieInformationDetailActivity
 import woowacourse.movie.list.adapter.TheaterAdapter
 import woowacourse.movie.list.model.TheaterData.theaters
+import woowacourse.movie.list.presenter.TheaterPresenter
 import woowacourse.movie.list.view.MovieListFragment.Companion.EXTRA_MOVIE_ID_KEY_TO_FRAGMENT
 
 class TheaterBottomSheetFragment : BottomSheetDialogFragment(), TheaterFragmentContract.View {
